feat: dependency 타입에 따른 gradle 파일 분리
