+++ app/build.gradle.kts
@@ -1,4 +1,5 @@
 plugins {
     id("base-conventions")
     id("kotlin-conventions")
+    id("spring-web-conventions")
 }
