+++ buildSrc/gradle.properties
@@ -1,0 +1,5 @@
+kotlin_gradle_plugin=2.1.0
+kotlin_version=2.1.0
+spring_boot_gradle_plugin=3.5.5
+spring_dependency_management_plugin=1.1.7
+jackson_module_version=2.15.2
