+++ buildSrc/src/main/kotlin/base-conventions.gradle.kts
@@ -1,0 +1,27 @@
+plugins {
+	id("org.jetbrains.kotlin.jvm")
+	id("org.jetbrains.kotlin.plugin.spring")
+	id("org.springframework.boot")
+	id("io.spring.dependency-management")
+}
+
+java {
+	toolchain {
+		languageVersion = JavaLanguageVersion.of(21)
+	}
+}
+
+repositories {
+	mavenCentral()
+}
+
+
+kotlin {
+	compilerOptions {
+		freeCompilerArgs.addAll("-Xjsr305=strict")
+	}
+}
+
+tasks.withType<Test> {
+	useJUnitPlatform()
+}
