+++ buildSrc/src/main/kotlin/kotlin-conventions.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+    id("org.jetbrains.kotlin.jvm")
+}
+
+dependencies {
+    implementation(kotlin("stdlib"))
+    implementation(kotlin("reflect"))
+    implementation("com.fasterxml.jackson.module:jackson-module-kotlin")
+}
