--- domain/src/main/kotlin/playground/cryptobackedloans/app/client/domain/loan/LoanStatus.kt
@@ -1,8 +1,0 @@
-package playground.cryptobackedloans.app.client.domain.loan
-
-enum class LoanStatus {
-    ACTIVE,
-    MARGIN_CALL,
-    LIQUIDATION,
-    REPAYMENT
-}
