--- domain/src/main/kotlin/playground/cryptobackedloans/app/client/domain/vo/Ltv.kt
@@ -1,5 +1,0 @@
-package playground.cryptobackedloans.app.client.domain.vo
-
-data class Ltv(
-    val ratio: Int
-)
