--- domain/src/main/kotlin/playground/cryptobackedloans/app/client/domain/vo/Rate.kt
@@ -1,5 +1,0 @@
-package playground.cryptobackedloans.app.client.domain.vo
-
-data class Rate(
-    val value: Int
-)
