+++ domain/src/main/kotlin/playground/cryptobackedloans/domain/loan/LiquidationType.kt
@@ -1,0 +1,9 @@
+package playground.cryptobackedloans.domain.loan
+
+enum class LiquidationType(
+    val description: String
+) {
+    UNPAID_INTEREST("이자 미납"),
+    UNPAID_PRINCIPAL("원금 미납"),
+    EXCEEDING_LTV("LTV 청산 조건 발동")
+}
