+++ domain/src/main/kotlin/playground/cryptobackedloans/domain/loan/LoanActivityHistory.kt
@@ -1,0 +1,14 @@
+package playground.cryptobackedloans.domain.loan
+
+import java.time.LocalDateTime
+
+/**
+ * 대출 이용 내역
+ */
+class LoanActivityHistory(
+    val id: Long,
+    val type: LoanActivityHistoryType,
+    val createdAt: LocalDateTime,
+    val updatedAt: LocalDateTime
+) {
+}
