+++ domain/src/main/kotlin/playground/cryptobackedloans/domain/loan/LoanStatus.kt
@@ -1,0 +1,8 @@
+package playground.cryptobackedloans.domain.loan
+
+enum class LoanStatus {
+    ACTIVE,
+    MARGIN_CALL,
+    LIQUIDATION,
+    REPAYMENT
+}
