+++ domain/src/main/kotlin/playground/cryptobackedloans/domain/loanProduct/Apr.kt
@@ -1,0 +1,18 @@
+package playground.cryptobackedloans.domain.loanProduct
+
+import playground.cryptobackedloans.domain.vo.Ltv
+import playground.cryptobackedloans.domain.vo.Rate
+import java.time.LocalDateTime
+
+/**
+ * 연 이율 (APR, Annual Percentage Rate)
+ */
+class Apr(
+    val id: Long,
+    val lowerBound: Ltv,
+    val upperBound: Ltv,
+    val rate: Rate,
+    val createdAt: LocalDateTime,
+    val updatedAt: LocalDateTime
+) {
+}
