+++ domain/src/main/kotlin/playground/cryptobackedloans/domain/vo/FiatPrice.kt
@@ -1,0 +1,8 @@
+package playground.cryptobackedloans.domain.vo
+
+import java.math.BigDecimal
+
+data class FiatPrice(
+    val value: BigDecimal
+) {
+}
