+++ domain/src/main/kotlin/playground/cryptobackedloans/domain/vo/Ltv.kt
@@ -1,0 +1,5 @@
+package playground.cryptobackedloans.domain.vo
+
+data class Ltv(
+    val ratio: Int
+)
