+++ domain/src/main/kotlin/playground/cryptobackedloans/domain/vo/Rate.kt
@@ -1,0 +1,5 @@
+package playground.cryptobackedloans.domain.vo
+
+data class Rate(
+    val value: Int
+)
